+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add support for the HINCRBYFLOAT command
 * Add support for `count` parameter on SRANDMEMBER
 * Add support for the PTTL command
+* Improve support for negative start indices in LRANGE
 
 ### 0.10.0
 * Add support for :nx, :xx, :ex, :px options for SET command